Migrate computation to Dask
Dask is used as computation backend. To simplify the API, a class QRCScheduler is added.
It takes as input a yaml config file where the regressors that we want to create and train
are specified. Internally, it creates a dictionary containing the instances of
quantileRegression_chain and quantileRegression_chain_disc. These are then passed to the Dask
scheduler which sends them to the workers.
Options to run both locally (default) and on SLURM (setup_slurm_cluster) are provided.